Added an explicit cast to suppress warnings
git-svn-id: http://svn.boost.org/svn/boost/trunk@12919 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/boost/type_traits/type_traits_test.hpp
@@ -42,7 +42,9 @@ int check_result(int argc, char** argv)
       std::cout << "Press any key to continue...";
       std::cin.get();
    }
-   return (failures == expected_failures) ? 0 : (failures != 0) ? failures : -1;
+   return (failures == expected_failures)
+       ? 0
+       : (failures != 0) ? static_cast<int>(failures) : -1;
 }
